+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 *   Added ability to set proxy at runtime (Dmitry Vorotilin)
 *   Implement basic support for accessing `SameSite` attribute of cookies (Reed Loden)
 *   Add clear_memory_cache to the driver (Piotr Gaertig)
+*   Allowing passing format and quality options to save_screenshot (Josef Šimánek, Thomas Walpole)
 
 #### Bug fixes ####
 *   Fix within_frame when called on a frame whose src attribute is about:blank (Thomas Walpole) [Issue #772]